Simplify assignment to __version__
This commit removes `get_version` helper which is completely useless and
not used anywhere else, i.e. from now on `__version__` is assigned directly.
As a side-effect, this makes patch number always visible in the
version number. This is correct behavior; there is no reason to hide it
when it equals 0.

Links: https://semver.org/