Remove content_type from Attribute class
Content-type field is used to limit Attributes that can be assigned to
Entity objects (see get_all_attributes method on Entity), and limit
Entities that can be assigned an Attribute. This way Attributes with
content_type assigned are, for all intended purposes, private to Entity
of that content_type. The problem is that those Attributes are still visible to
other Entities. Moreover, originally, attributes had
non-unique slugs that are unique together with content_type and site.
With removal of site (and because content_type is nullable) a situation
might exist in which two Attributes exist with the same slug/name but
one with content_type and one with NULL. Simple solution to all
this hussle is to drop private Attributes alltogether; we believe this
complication is just not worth it.